small refactoring, conventional POSIX-compatible cli lib, convinient log library, systemd and docker configs
